Function alias_thing::utils::get_shell::get_shell

pub fn get_shell() -> Result<String, WorkspaceError>

Retrieves the shell configuration file path based on the value of the SHELL environment variable.

This function attempts to retrieve the shell configuration file path based on the value of the SHELL environment variable. If the SHELL environment variable is set and contains the string “zsh”, it returns ".zshrc". Otherwise, it returns ".bashrc".

If the SHELL environment variable is not set or if an error occurs while accessing it, an error of type WorkspaceError is returned.

Errors

This function can return an error of type WorkspaceError if:

The SHELL environment variable is not set or an error occurs while accessing it.

Examples

use crate::get_shell;

match get_shell() {
    Ok(shell_path) => println!("Shell configuration file path: {}", shell_path),
    Err(err) => eprintln!("Error: {:?}", err),
}